+++ workspaces/frontend/src/__mocks__/mockNamespaces.ts
@@ -1,5 +1,5 @@
 import { buildMockNamespace } from '~/shared/mock/mockBuilder';
-import { Namespace } from '~/shared/types';
+import { Namespace } from '~/shared/api/backendApiTypes';
 
 export const mockNamespaces: Namespace[] = [
   buildMockNamespace({ name: 'default' }),
