+++ workspaces/frontend/src/__mocks__/utils.ts
@@ -1,4 +1,4 @@
-import { ResponseBody } from '~/app/types';
+import { ResponseBody } from '~/shared/api/types';
 
 export const mockBFFResponse = <T>(data: T): ResponseBody<T> => ({
   data,
